Code cleanup. Changed to utilizing JRE tar.gz releases directly rather than an unpacked tar.gz. Changed and improved specification of JRE to allow for better flexibility.

+++ script/appleUtils.py
@@ -1,14 +1,10 @@
 #! /usr/bin/env python
 
-import argparse
 import copy
-import getpass
 import math
 import os
 import shutil
-import signal
 import subprocess
-import sys
 import tempfile
 import time
 import glob
@@ -18,71 +14,78 @@ import miscUtils
 
 
 def buildRelease(args, buildPath):
-	# We may mutate args - thus make a custom copy  
+	# We mutate args - thus make a custom copy
 	args = copy.copy(args)
-	
+
 	# Retrieve vars of interest
 	appName = args.name
 	version = args.version
 	jreRelease = args.jreRelease
-	
-	# Attempt to locate a default JRE if none is specified in the args.
+	platformStr = 'apple'
+
+	# Check our system environment before proceeding
+	if checkSystemEnvironment() == False:
+		return
+
+	# Attempt to locate a default JRE if None is specified in the args.
 	if jreRelease == None:
-		jreRelease = jreUtils.getDefaultJreRelease('apple')
+		jreRelease = jreUtils.getDefaultJreRelease(platformStr)
 	# Let the user know if the 'user' specified JRE is not available and locate an alternative
-	elif jreUtils.isJreAvailable('apple', jreRelease) == False:
-		print('[Warning] User specified JRE (' + jreRelease + ') is not available for Apple platform. Searching for alternative...')
-		jreRelease = jreUtils.getDefaultJreRelease('apple')
+	elif jreUtils.getJreTarGzFile(platformStr, jreRelease) == None:
+		print('[Warning] User specified JRE ({0}) is not available for {1} platform. Searching for alternative...'.format(jreRelease, platformStr.capitalize()))
+		jreRelease = jreUtils.getDefaultJreRelease(platformStr)
 	args.jreRelease = jreRelease
 
-	# Form the list of distributions to build
-	distList = [(appName + '-' + version, False)]
-	if jreUtils.isJreAvailable('apple', jreRelease) == True:
-		distList.append((appName + '-' + version + '-jre', True))
+	# Form the list of distributions to build (dynamic and static JREs)
+	distList = [(appName + '-' + version, None)]
+	jreTarGzFile = jreUtils.getJreTarGzFile(platformStr, jreRelease)
+	if jreTarGzFile != None:
+		distList.append((appName + '-' + version + '-jre', jreTarGzFile))
 
-	# Create the various Apple distributions	
-	for (distName, isStaticRelease) in distList:
-		print('Building Apple distribution: ' + distName)
-		# Let the user know of the JRE release we are going to build with
-		if isStaticRelease == True:
-			print('\tUtilizing jreRelease: ' + jreRelease)
-		
-		# Create a tmp folder and build the static release to the tmp folder 
-		tmpPath = tempfile.mkdtemp(dir=buildPath)
-		
+	# Create the various distributions
+	for (aDistName, aJreTarGzFile) in distList:
+		print('Building {0} distribution: {1}'.format(platformStr.capitalize(), aDistName))
+		# Let the user know of the JRE tar.gz  we are going to build with
+		if aJreTarGzFile != None:
+			print('\tUtilizing jreRelease: ' + aJreTarGzFile)
+
+		# Create a tmp folder and build the static release to the tmp folder
+		tmpPath = tempfile.mkdtemp(prefix=platformStr, dir=buildPath)
+
 		# Build the contents of the distribution folder
-		buildDistTree(buildPath, tmpPath, args, isStaticRelease)
-		
+		buildDistTree(buildPath, tmpPath, args, aJreTarGzFile)
+
 		# Create the DMG image via genisoimage
-		dmgFile = os.path.join(buildPath, distName + '.dmg')
+		dmgFile = os.path.join(buildPath, aDistName + '.dmg')
 #		cmd = ['genisoimage', '-o', dmgFile, '-quiet', '-V', appName, '-max-iso9660-filenames', '-hfs-unlock', '-uid', '501', '-gid', '80', '-r', '-D', '-apple', tmpPath]
 		cmd = ['genisoimage', '-o', dmgFile, '-quiet', '-V', appName, '-max-iso9660-filenames', '-hfs-unlock', '-D', '-r', '-apple', tmpPath]
 		print('\tForming DMG image. File: ' + dmgFile)
 		proc = miscUtils.executeAndLog(cmd, "\t\tgenisoimage: ")
 		if proc.returncode != 0:
 			print('\tError: Failed to form DMG image. Return code: ' + str(proc.returncode))
+		else:
+			print('\tFinished building release: ' + os.path.basename(dmgFile))
 
-		# Perform cleanup: Remove the tmp folder
+		# Perform cleanup
 		shutil.rmtree(tmpPath)
-		print('\tFinished building release: ' + os.path.basename(dmgFile))
-	
-	
-	
-	
+
+
+
+
 #	# Compute the space needed for the entire dmg package
 #	fileSize = miscUtils.getPathSize(javaCodePath)
 #	fileSize += miscUtils.getPathSize(args.bgFile)
 #	fileSize += miscUtils.getPathSize(args.icnsFile)
 #	# Add extra space for Generic.app template and nominal Info.plist
 #	fileSize + 100 * 1024;
-	
+
 #	# Determine the 1 megabyte blocks that we need
 #	numBlocks = math.ceil(fileSize / (1024 * 1024.0));
 #	numBlocks = int(numBlocks)
-	
+
 #	# Extra padding in case files take up slightly more space
-#	numBlocks += 1;  
-	
+#	numBlocks += 1;
+
 #	# Create the empty file which will be used for the dmg file
 #	# dd if=/dev/zero of=FireFoxRedo.dmg bs=1M count=128
 #	print('Creating a ' + str(numBlocks) + ' MB file: ' + dmgFile)
@@ -95,21 +98,21 @@ def buildRelease(args, buildPath):
 #	exeTool = os.path.join(appInstallRoot, 'tools', 'mkfs.hfsplus')
 #	cmd = [exeTool, '-v', appName, dmgFile]
 #	subprocess.call(cmd, stderr=subprocess.STDOUT)
-	
-	
+
+
 #	# Mount the dmgFile
 #	mount(dmgFile, tmpPath)
-	
+
 #	# Build the contents of the DMG file
 #	buildSharedTree(tmpPath, args)
-	
-	
+
+
 #	# Change the group/user info to Apple specific guid
 #	normGuid(tmpPath)
 
 #	# Unmount the dmgFile
 #	umount(tmpPath)
-	
+
 #	# Let the umount command finish - Bail after 10 seconds
 #	for c1 in range(40):
 #		if os.path.ismount(tmpPath) == False:
@@ -123,9 +126,9 @@ def buildRelease(args, buildPath):
 
 #	# Perform cleanup: Remove the tmp folder
 #	os.rmdir(tmpPath)
-	
 
-def buildDistTree(buildPath, rootPath, args, isStaticRelease):
+
+def buildDistTree(buildPath, rootPath, args, jreTarGzFile):
 	# Retrieve vars of interest
 	appInstallRoot = miscUtils.getInstallRoot()
 	appInstallRoot = os.path.dirname(appInstallRoot)
@@ -134,69 +137,66 @@ def buildDistTree(buildPath, rootPath, args, isStaticRelease):
 	bgFile = args.bgFile
 	icnsFile = args.icnsFile
 	jreRelease = args.jreRelease
-	
-	# Form the symbolic link which points to /Applications 
+
+	# Form the symbolic link which points to /Applications
 	srcPath = '/Applications'
 	dstPath = os.path.join(rootPath, 'Applications');
 	os.symlink(srcPath, dstPath)
-	
+
 	# Construct the app folder
 	appNodes = ['MacOS', 'Resources']
-	if isStaticRelease == True:
+	if jreTarGzFile != None:
 		appNodes.append('PlugIns')
 	for aPath in appNodes:
 		dstPath = os.path.join(rootPath, appName + '.app', 'Contents', aPath)
 		os.makedirs(dstPath)
-		
-	if isStaticRelease == True:
+
+	if jreTarGzFile != None:
 		# Copy over the executable launcher
 		srcPath = os.path.join(appTemplatePath, 'apple', 'JavaAppLauncher')
 		dstPath = os.path.join(rootPath, appName + '.app', 'Contents', 'MacOS')
 		shutil.copy(srcPath, dstPath)
-	
-		# Copy over the JRE tree
-		srcPath = os.path.join(appInstallRoot, 'jre', 'apple', jreRelease)
-		dstPath = os.path.join(rootPath, appName + '.app', 'Contents', 'PlugIns', jreRelease)
-		shutil.copytree(srcPath, dstPath, symlinks=True)
+
+		# Unpack the JRE and set up the JRE tree
+		destPath = os.path.join(rootPath, appName + '.app', 'Contents', 'PlugIns')
+		jreUtils.unpackAndRenameToStandard(jreTarGzFile, destPath)
 	else:
 		# Copy over the executable launcher
 		srcPath = os.path.join(appTemplatePath, 'apple', 'JavaApplicationStub')
 		dstPath = os.path.join(rootPath, appName + '.app', 'Contents', 'MacOS')
 		shutil.copy(srcPath, dstPath)
-	
-	# Write out the PkgInfo file	
+
+	# Write out the PkgInfo file
 	dstPath = os.path.join(rootPath, appName + '.app', 'Contents', "PkgInfo")
 	f = open(dstPath, 'wb')
 	f.write('APPL????')
 	f.close()
-	
+
 	# Determine the payloadPath for where to store the appLauncher
 	payloadPath = os.path.join(rootPath, appName + '.app', 'Contents')
-	if isStaticRelease == False:
+	if jreTarGzFile == None:
 		payloadPath = os.path.join(rootPath, appName + '.app', 'Contents', 'Resources')
-	
+
 	# Form the app contents folder
 	srcPath = os.path.join(buildPath, "delta")
 	dstPath = os.path.join(payloadPath, 'app')
 	shutil.copytree(srcPath, dstPath, symlinks=True)
 
-	#Link dlls to the MacOS directory so they can be found at launch
-	jarDir = os.path.join(rootPath, appName + '.app', 'Contents', 'app', 'code','osx')
+	# Link dlls to the MacOS directory so they can be found at launch
+	jarDir = os.path.join(rootPath, appName + '.app', 'Contents', 'app', 'code', 'osx')
 	dstPath = os.path.join(rootPath, appName + '.app', 'Contents', 'MacOS')
-	for jniPath in glob.iglob(os.path.join(jarDir,"*.jnilib")):
+	for jniPath in glob.iglob(os.path.join(jarDir, "*.jnilib")):
 		jniFileName = os.path.basename(jniPath)
-		srcPath = os.path.join('..','app','code','osx',jniFileName)
-		linkPath = os.path.join(dstPath,jniFileName)
-		os.symlink(srcPath,linkPath)
-	for dylPath in glob.iglob(os.path.join(jarDir,"*.dylib")):
+		srcPath = os.path.join('..', 'app', 'code', 'osx', jniFileName)
+		linkPath = os.path.join(dstPath, jniFileName)
+		os.symlink(srcPath, linkPath)
+	for dylPath in glob.iglob(os.path.join(jarDir, "*.dylib")):
 		dylFileName = os.path.basename(dylPath)
-		srcPath = os.path.join('..','app','code','osx',dylFileName)
-		linkPath = os.path.join(dstPath,dylFileName)
-		os.symlink(srcPath,linkPath)
+		srcPath = os.path.join('..', 'app', 'code', 'osx', dylFileName)
+		linkPath = os.path.join(dstPath, dylFileName)
+		os.symlink(srcPath, linkPath)
 
 
-
-	
 	# Setup the launcher contents
 	exePath = os.path.join(payloadPath, 'Java')
 	srcPath = os.path.join(appInstallRoot, "template/appLauncher.jar")
@@ -207,17 +207,17 @@ def buildDistTree(buildPath, rootPath, args, isStaticRelease):
 	if args.javaCode != None:
 		# Form the Info.plist file
 		dstPath = os.path.join(rootPath, appName + '.app', 'Contents', 'Info.plist')
-		if isStaticRelease == True:
+		if jreTarGzFile != None:
 			buildPListInfoStatic(dstPath, args)
 		else:
-			buildPListInfoShared(dstPath, args)		
-	
+			buildPListInfoShared(dstPath, args)
+
 	# Copy over the icon file *.icns
-	if icnsFile != None and os.path.exists(icnsFile) == True: 
+	if icnsFile != None and os.path.exists(icnsFile) == True:
 		srcPath = icnsFile
 		dstPath = os.path.join(rootPath, appName + '.app', 'Contents', 'Resources')
 		shutil.copy(srcPath, dstPath)
-	
+
 	# Copy over the background file
 	srcPath = bgFile
 	if srcPath == None:
@@ -226,12 +226,12 @@ def buildDistTree(buildPath, rootPath, args, isStaticRelease):
 	os.mkdir(dstPath)
 	dstPath = os.path.join(rootPath, '.background', 'background.png')
 	shutil.copy(srcPath, dstPath)
-	
+
 	# Copy over the .DS_Store
 	srcPath = os.path.join(appTemplatePath, '.DS_Store.template')
 	dstPath = os.path.join(rootPath, '.DS_Store')
 	shutil.copy(srcPath, dstPath)
-	
+
 	# Update the .DS_Store file to reflect the new volume name
 	srcPath = os.path.join(rootPath, '.DS_Store')
 	classPath = appInstallRoot + '/lib/glum.jar:' + appInstallRoot + '/lib/distMaker.jar:' + appInstallRoot + '/lib/guava-18.0.jar'
@@ -240,17 +240,17 @@ def buildDistTree(buildPath, rootPath, args, isStaticRelease):
 	if proc.returncode != 0:
 		print('\tError: Failed to update .DS_Store. Return code: ' + str(proc.returncode))
 
-	
 
 
 
 
 
-	
+
+
 def normGuid(mountPt):
 	# Ensure we are running as root
 	miscUtils.checkRoot()
-	
+
 	# Ensure mountPt is a loopback mountPt
 	# TODO
 
@@ -265,41 +265,41 @@ def normGuid(mountPt):
 #		for f in files:
 #			tmpPath = '\'' + os.path.join(path, f) + '\''
 #			os.chown(tmpPath, 501, 80)
-	
+
 
 
 def mount(dmgFile, mountPt):
 	# Ensure we are running as root
 	miscUtils.checkRoot()
-	
+
 	# Mount the dmgFile (silently ???)
 #	mount -o loop,ro -t hfsplus imagefile.dmg /mnt/mountpoint
 	cmd = ['mount', '-n', '-o', 'loop,rw', '-t', 'hfsplus', dmgFile, mountPt]
 	subprocess.call(cmd, stderr=subprocess.STDOUT)
-	
+
 
 def umount(mountPt):
 	# Ensure we are running as root
 	miscUtils.checkRoot()
 
-	# Release the mount	
+	# Release the mount
 	cmd = ['umount', mountPt]
 	subprocess.call(cmd, stderr=subprocess.STDOUT)
-	
+
 
 def buildPListInfoShared(destFile, args):
 	# Retrieve vars of interest
 	icnsStr = None
 	if args.icnsFile != None:
 		icnsStr = os.path.basename(args.icnsFile)
-	
+
 	classPathStr = '$JAVAROOT/appLauncher.jar'
 #	classPathStr = ''
 #	for aStr in args.classPath:
 #		classPathStr += '$JAVAROOT/' + aStr + ':'
 #	if len(classPathStr) > 0:
 #		classPathStr = classPathStr[0:-1]
-		
+
 	jvmArgsStr = ''
 	for aStr in args.jvmArgs:
 		jvmArgsStr += aStr + ' '
@@ -310,7 +310,7 @@ def buildPListInfoShared(destFile, args):
 #	writeln(f, 0, '<!DOCTYPE plist PUBLIC "-//Apple Computer//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">')
 	writeln(f, 0, '<plist version="1.0">')
 	writeln(f, 1, '<dict>')
-	
+
 	tupList = []
 	tupList.append(('CFBundleExecutable', 'JavaApplicationStub'))
 	tupList.append(('CFBundleGetInfoString', args.company))
@@ -321,24 +321,24 @@ def buildPListInfoShared(destFile, args):
 	tupList.append(('CFBundlePackageType', 'APPL'))
 	tupList.append(('CFBundleSignature', '????'))
 	tupList.append(('CFBundleVersion', args.version))
-	
+
 	# Application configuration
-	for (key,val) in tupList:
+	for (key, val) in tupList:
 		writeln(f, 2, '<key>' + key + '</key>')
 		writeln(f, 2, '<string>' + str(val) + '</string>')
-	
+
 	# JVM configuration
 	writeln(f, 2, '<key>Java</key>')
 	writeln(f, 2, '<dict>')
-	
+
 	tupList = []
 	tupList.append(('JVMVersion', '1.7+'))
 	tupList.append(('MainClass', 'appLauncher.AppLauncher'))
 	tupList.append(('WorkingDirectory', '$APP_PACKAGE/Contents/Resources/app'))
 	tupList.append(('ClassPath', classPathStr))
 	tupList.append(('VMOptions', jvmArgsStr))
-	
-	for (key,val) in tupList:
+
+	for (key, val) in tupList:
 		writeln(f, 3, '<key>' + key + '</key>')
 		writeln(f, 3, '<string>' + str(val) + '</string>')
 
@@ -350,7 +350,7 @@ def buildPListInfoShared(destFile, args):
 	writeln(f, 2, '</dict>')
 	writeln(f, 1, '</dict>')
 	writeln(f, 0, '</plist>')
-	
+
 	f.close()
 
 def buildPListInfoStatic(destFile, args):
@@ -358,20 +358,20 @@ def buildPListInfoStatic(destFile, args):
 	icnsStr = None
 	if args.icnsFile != None:
 		icnsStr = os.path.basename(args.icnsFile)
-	
+
 	classPathStr = '$JAVAROOT/appLauncher.jar'
 #	classPathStr = ''
 #	for aStr in args.classPath:
 #		classPathStr += '$JAVAROOT/' + aStr + ':'
 #	if len(classPathStr) > 0:
 #		classPathStr = classPathStr[0:-1]
-		
+
 	f = open(destFile, 'wb')
 	writeln(f, 0, '<?xml version="1.0" ?>')
 #	writeln(f, 0, '<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">')
 	writeln(f, 0, '<plist version="1.0">')
 	writeln(f, 1, '<dict>')
-	
+
 	tupList = []
 	tupList.append(('CFBundleDevelopmentRegion', 'English'))
 	tupList.append(('CFBundleExecutable', 'JavaAppLauncher'))
@@ -385,16 +385,17 @@ def buildPListInfoStatic(destFile, args):
 	tupList.append(('CFBundleSignature', '????'))
 	tupList.append(('CFBundleVersion', args.version))
 	tupList.append(('NSHumanReadableCopyright', ''))
-	tupList.append(('JVMRuntime', args.jreRelease))
-	
+	jrePath = 'jre' + args.jreRelease
+	tupList.append(('JVMRuntime', jrePath))
+
 	tupList.append(('JVMMainClassName', 'appLauncher.AppLauncher'))
-	
-	
+
+
 	# Application configuration
-	for (key,val) in tupList:
+	for (key, val) in tupList:
 		writeln(f, 2, '<key>' + key + '</key>')
 		writeln(f, 2, '<string>' + str(val) + '</string>')
-		
+
 	# JVM options
 	writeln(f, 2, '<key>JVMOptions</key>')
 	writeln(f, 2, '<array>')
@@ -409,19 +410,19 @@ def buildPListInfoStatic(destFile, args):
 	writeln(f, 3, '<string>-Dapple.awt.application.name=' + args.name + '</string>')
 	writeln(f, 3, '<string>-Djava.system.class.loader=appLauncher.RootClassLoader</string>')
 	writeln(f, 2, '</array>')
-	
+
 #	# App arguments
 #	writeln(f, 2, '<key>JVMArguments</key>')
 #	writeln(f, 3, '<array>')
 #	for aStr in args.appArgs:
 #		writeln(f, 4, '<string>' + args + '</string>')
 #	writeln(f, 3, '</array>')
-	
+
 
 	# JVM configuration
 	writeln(f, 2, '<key>Java</key>')
 	writeln(f, 2, '<dict>')
-	
+
 	tupList = []
 #	tupList.append(('JVMVersion', '1.6+'))
 #	tupList.append(('MainClass', args.mainClass))
@@ -429,16 +430,21 @@ def buildPListInfoStatic(destFile, args):
 
 	tupList.append(('ClassPath', classPathStr))
 
-	for (key,val) in tupList:
+	for (key, val) in tupList:
 		writeln(f, 3, '<key>' + key + '</key>')
 		writeln(f, 3, '<string>' + str(val) + '</string>')
-	
+
 	writeln(f, 2, '</dict>')
 	writeln(f, 1, '</dict>')
 	writeln(f, 0, '</plist>')
-	
+
 	f.close()
 
+
+def checkSystemEnvironment():
+	"""Checks to ensure that all system application / environment variables needed to build a Windows distribution are installed
+	and properly configured. Returns False if the system environment is insufficient"""
+	return True
 
 
 def writeln(f, tabL, aStr, tabStr='   '):
@@ -446,5 +452,4 @@ def writeln(f, tabL, aStr, tabStr='   '):
 	for i in range(tabL):
 		tStr += tabStr
 	f.write(tStr + aStr + '\n')
-	
-	
+
